+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+## v41 (11/1/2012)
+
+Features:
+
+* Enable ruby 2.0.0 support for testing
+
+## v40 (10/14/2012)
+
+Features:
+
+* Cache version of the buildpack we used to deploy
+* Purge cache when v38 is detected
+
+## v39 (10/14/2012)
+
+Bugfixes:
+
+* Don't display cache clearing message for new apps
+* Actually clear bundler cache on ruby version change
+
+## v38 (10/14/2012)
+
+Bugfixes:
+
+* Stop bundle cache from continually growing
+
+## v37 (10/12/2012)
+
+Bugfixes:
+
+* Remove temporary workaround from v36.
+* Clear bundler cache upon Ruby version change
+
+## v36 (10/12/2012)
+
+Bugfixes:
+
+* Always clear the cache for ruby 1.9.3 as a temporary workaround due to the security upgrade
+
 ## v35 (9/19/2012)
 
 Features: